Use Py_ssize_t for loop counter in wsgi_convert_headers_to_bytes.
Matches the type returned by PyList_Size for consistency with
similar cleanups done elsewhere in the code base.

+++ src/server/wsgi_convert.c
@@ -70,7 +70,7 @@ PyObject *wsgi_convert_headers_to_bytes(PyObject *headers)
 {
     PyObject *result = NULL;
 
-    int i;
+    Py_ssize_t i;
     Py_ssize_t size;
 
     if (!PyList_Check(headers))
